Option to clamp fractional zoom to either nearest or higher zoom
This is a remake of PR #6034 by @boundaryfree. At fractional zoom levels, tiles have to be either scaled up or down. Currently, Leaflet scales to the nearest integer zoom which means that the tiles can be scaled up leading to blurry images. This is particularly important for non-geographic images, e.g., when Leaflet is used to display pyramids with the Zoomify or Deep Zoom plugins - the blurring is quite noticeable for high-quality images such as photographs! It can also important for vector drawings such as CAD drawings.

This PR adds an option for the user to specify whether to scale the tiles at fractional zoom to the nearest integer zoom (default), or always scale them down from the higher integer zoom (this gives sharper images but increases download size).

The issue of blurred tiles in Leaflet has been discussed a few times before, e.g.: 
https://stackoverflow.com/questions/40387914/force-leaflet-map-to-use-only-integer-zoom-levels-no-fractional-levels
https://github.com/Leaflet/Leaflet/issues/6068
https://github.com/Leaflet/Leaflet/issues/6069

Here is a relevant detailed discussion of the same issue in mapbox:
https://github.com/mapbox/mapbox-gl-js/issues/1030